% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corzed-package.R
\docType{package}
\name{corzed-package}
\alias{corzed-package}
\title{corzed: Location-adjusted Wald statistic for general model classes}
\description{
ADD PACKAGE DESCRIPTION HERE
}
\details{
ANY OTHER DETAILS?
}
\references{
Di Caterina C, Kosmidis I (2017). Location-adjusted Wald statistic
for scalar parameters *arXiv*, **arXiv:1710.11217**
}
\author{
Ioannis Kosmidis \email{i.kosmidis@ucl.ac.uk}, Claudia Di Caterina \email{dicaterina@stats.unipd.it}
}
