% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corzed_old.R
\name{corzed.glm}
\alias{corzed.glm}
\title{Location-adjusted Wald statistics}
\usage{
\method{corzed}{glm}(object, null = 0, correction = TRUE, ncores = 1,
  what = NULL, analytical = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{\link{glm}}}

\item{null}{a numeric vector of the individual null hypotheses to
be tested. Default is that each parameter is zero}

\item{correction}{should the correction be applied? Default is
\code{TRUE}. If \code{TRUE} then the location-adjusted z
statistic is computed}

\item{ncores}{how many computing nodes should be used for the
calculation of the adjusted Wald statistics (one node per
parameter). Default is 1}

\item{...}{placeholder for extra arguments. Currently not used}
}
\description{
Location-adjusted Wald statistics
}
\examples{
clotting <- data.frame(
   conc = c(118,58,42,35,27,25,21,19,18,69,35,26,21,18,16,13,12,12),
   u = c(5,10,15,20,30,40,60,80,100, 5,10,15,20,30,40,60,80,100),
   lot = factor(c(rep(1, 9), rep(2, 9))))

modML <- glm(conc ~ log(u)*lot, data = clotting, family = Gamma(link="log"))
corzed(modML)

## Aliasing
modMLo <- glm(conc ~ log(u)*lot + I(2*log(u)), data = clotting, family = Gamma(link="log"))
corzed(modMLo)

}
\seealso{
\code{\link{summary.glm}}
}
\author{
Ioannis Kosmidis [aut, cre] \email{i.kosmidis@ucl.ac.uk},
    Claudia Di Caterina [ctb] \email{dicaterina@stat.unipd.it}
}
