% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corzed-package.R
\name{corzed}
\alias{corzed}
\title{Generic method for computing location-adjusted Wald statistics for
the parameters of general models}
\usage{
corzed(object, null = 0, adjust = TRUE, which = NULL, parallel = FALSE,
  ...)
}
\arguments{
\item{object}{a fitted model object (e.g. the result of a
\code{\link{glm}} call)}

\item{null}{a scalar or a vector of the same length as
\code{coef(object)} defining the parameter values for the null
hypotheses. \code{0} by default}

\item{adjust}{if \code{TRUE} (default) compute the
location-adjusted Wald statistic is returned, if \code{FALSE}
return the standard Wald statistic}

\item{which}{for which parameters should the statistics be
computed? This can be a numeric or character vector.  The
default (\code{NULL}) is to compute statistics for all
parameters}

\item{parallel}{if \code{TRUE}, the statistics are computed in
parallel using the backend provided to \code{\link{foreach}}}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Generic method for computing location-adjusted Wald statistics for
the parameters of general models
}
\details{
\code{corzed} stands for "cor"rected "zed"-statistic.
}
\references{
Di Caterina C, Kosmidis I (2017). Location-adjusted Wald statistic
for scalar parameters *arXiv*, **arXiv:1710.11217**
}
\seealso{
\code{\link{corzed.glm}}
}
\author{
Ioannis Kosmidis \email{i.kosmidis@ucl.ac.uk},
        Claudia Di Caterina \email{dicaterina@stat.unipd.it}
}
