% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldi_betareg.R
\name{waldi.betareg}
\alias{waldi.betareg}
\title{Location-adjusted Wald statistics for \code{betareg} objects}
\usage{
\method{waldi}{betareg}(object, null = 0, adjust = TRUE,
  which = NULL, parallel = FALSE, numerical = TRUE, ...)
}
\arguments{
\item{object}{a fitted model object (e.g. the result of a
\code{\link{glm}} call)}

\item{null}{a scalar or a vector of the same length as
\code{coef(object)} defining the parameter values for the null
hypotheses. \code{0} by default}

\item{adjust}{if \code{TRUE} (default) return the location-adjusted
Wald statistic, if \code{FALSE} return the standard Wald
statistic}

\item{which}{for which parameters should the statistics be
computed? This can be a numeric or character vector.  The
default (\code{NULL}) is to compute statistics for all
parameters}

\item{parallel}{if \code{TRUE}, the statistics are computed in
parallel using the backend provided to \code{\link{foreach}}}

\item{numerical}{shall numerical derivatives be used for the
computation of the location-adjusted statistics? Default is \code{TRUE}. Currently unused.}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Location-adjusted Wald statistics for \code{betareg} objects
}
