% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldi_glm.R
\name{waldi.glm}
\alias{waldi.glm}
\title{Location-adjusted Wald statistics for \code{\link{glm}} and
\code{\link[brglm2]{brglmFit}} objects}
\usage{
\method{waldi}{glm}(object, null = 0, adjust = TRUE, which = NULL,
  parallel = FALSE, numerical = TRUE, ...)
}
\arguments{
\item{object}{a fitted model object (e.g. the result of a
\code{\link{glm}} call)}

\item{null}{a scalar or a vector of the same length as
\code{coef(object)} defining the parameter values for the null
hypotheses. \code{0} by default}

\item{adjust}{if \code{TRUE} (default) return the location-adjusted
Wald statistic, if \code{FALSE} return the standard Wald
statistic}

\item{which}{for which parameters should the statistics be
computed? This can be a numeric or character vector.  The
default (\code{NULL}) is to compute statistics for all
parameters}

\item{parallel}{if \code{TRUE}, the statistics are computed in
parallel using the backend provided to \code{\link{foreach}}}

\item{numerical}{shall numerical derivatives be used for the
computation of the location-adjusted statistics? Default is \code{TRUE}}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Location-adjusted Wald statistics for \code{\link{glm}} and
\code{\link[brglm2]{brglmFit}} objects
}
\examples{
clotting <- data.frame(
   conc = c(118,58,42,35,27,25,21,19,18,69,35,26,21,18,16,13,12,12),
   u = c(5,10,15,20,30,40,60,80,100, 5,10,15,20,30,40,60,80,100),
   lot = factor(c(rep(1, 9), rep(2, 9))))

modML <- glm(conc ~ log(u)*lot, data = clotting, family = Gamma(link="log"))
waldi(modML, parallel = FALSE)

## Unidentifiability
modMLo <- glm(conc ~ log(u)*lot + I(2*log(u)), data = clotting, family = Gamma(link="log"))
waldi(modMLo)

}
\seealso{
\code{\link{summary.glm}}
}
