% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldi-package.R
\name{waldi}
\alias{waldi}
\title{Generic method for computing location-adjusted Wald statistics for
the parameters of general models}
\usage{
waldi(object, null = 0, adjust = TRUE, which = NULL, parallel = FALSE,
  ...)
}
\arguments{
\item{object}{a fitted model object (e.g. the result of a
\code{\link{glm}} call)}

\item{null}{a scalar or a vector of the same length as
\code{coef(object)} defining the parameter values for the null
hypotheses. \code{0} by default}

\item{adjust}{if \code{TRUE} (default) return the location-adjusted
Wald statistic, if \code{FALSE} return the standard Wald
statistic}

\item{which}{for which parameters should the statistics be
computed? This can be a numeric or character vector.  The
default (\code{NULL}) is to compute statistics for all
parameters}

\item{parallel}{if \code{TRUE}, the statistics are computed in
parallel using the backend provided to \code{\link{foreach}}}

\item{...}{further arguments to be passed to or from other methods}
}
\description{
Generic method for computing location-adjusted Wald statistics for
the parameters of general models
}
\details{
ADD TECHNICAL DETAILS HERE
}
\references{
Di Caterina C, Kosmidis I (2017). Location-adjusted Wald statistic
for scalar parameters *arXiv*, **arXiv:1710.11217**
}
\seealso{
\code{\link{waldi.glm}} \code{\link{waldi.betareg}}
}
