% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldi-package.R
\docType{package}
\name{waldi-package}
\alias{waldi-package}
\title{waldi: Location-adjusted Wald statistic for general model classes}
\description{
ADD PACKAGE DESCRIPTION HERE
}
\details{
ANY OTHER DETAILS?
}
\references{
Di Caterina C, Kosmidis I (2017). Location-adjusted Wald statistic
for scalar parameters *arXiv*, **arXiv:1710.11217**
}
\author{
Ioannis Kosmidis \email{i.kosmidis@ucl.ac.uk},
        Claudia Di Caterina \email{dicaterina@stat.unipd.it}
}
